0-data-preparation/facilities/requirements.txt
osmnx==1.0.1
plotly==4.14.3
tqdm==4.60.0
matsim-tools==0.0.10